+++ wagtail_newsletter/actions.py
@@ -20,6 +20,8 @@ def save_campaign(request, page: NewsletterPageMixin) -> None:
             recipients=version.newsletter_recipients,
             subject=subject,
             html=version.get_newsletter_html(),
+            from_name=version.get_newsletter_from_name(),
+            reply_to=version.get_newsletter_reply_to(),
         )
 
     except campaign_backends.CampaignBackendError as error:
@@ -98,9 +100,15 @@ def schedule_campaign(request, page: NewsletterPageMixin) -> None:
 
     schedule_time = form.cleaned_data["schedule_time"]
 
-    save_campaign(request, page)
-
     backend = campaign_backends.get_backend()
+
+    try:
+        backend.validate_schedule_time(schedule_time)
+    except campaign_backends.CampaignBackendError as error:
+        messages.error(request, error.message)
+        return
+
+    save_campaign(request, page)
 
     try:
         backend.schedule_campaign(