+++ wagtail_newsletter/campaign_backends/mailchimp.py
@@ -20,13 +20,22 @@ from ..models import NewsletterRecipientsBase
 from . import Campaign, CampaignBackend, CampaignBackendError
 
 
+logger = logging.getLogger(__name__)
+
+
 class CampaignStatus(Enum):
     DRAFT = "save"
     SCHEDULED = "schedule"
     PAUSED = "paused"
 
 
-logger = logging.getLogger(__name__)
+class SegmentType(Enum):
+    STATIC = "static"  # Mailchimp Tags
+    SAVED = "saved"  # Mailchimp Segments created manually
+    CAMPAIGN_STATIC = (
+        "campaign_static"  # Pastes of email addresses into the campaign editor
+    )
+    FUZZY = "fuzzy"  # purpose remains fuzzy
 
 
 @dataclass
@@ -111,9 +120,22 @@ class MailchimpCampaignBackend(CampaignBackend):
             for audience in audiences
         ]
 
-    def get_audience_segments(self, audience_id) -> "list[AudienceSegment]":
+    def get_audience_segments(
+        self, audience_id, type=SegmentType.SAVED.value, count=1000
+    ) -> "list[AudienceSegment]":
+        """
+        Fetch audience segments from Mailchimp.
+
+        `type`: filter by segment type.  By default, only "saved" segments are fetched,
+        which are the segments created manually by the Mailchimp admin.
+
+        `count`: how many items to fetch. By default, 1000 items are fetched, which is
+        the maximum allowed by Mailchimp, so that we don't have to implement pagination.
+        """
         try:
-            segments = self.client.lists.list_segments(audience_id)["segments"]
+            segments = self.client.lists.list_segments(
+                audience_id, type=type, count=count
+            )["segments"]
 
         except ApiClientError as error:
             if error.status_code == 404:
@@ -170,26 +192,29 @@ class MailchimpCampaignBackend(CampaignBackend):
         self,
         *,
         recipients: Optional[NewsletterRecipientsBase],
+        from_name: str,
+        reply_to: str,
         subject: str,
     ) -> "dict[str, Any]":
         body = {
             "settings": {
-                "from_name": _require_setting("WAGTAIL_NEWSLETTER_FROM_NAME"),
-                "reply_to": _require_setting("WAGTAIL_NEWSLETTER_REPLY_TO"),
+                "from_name": from_name,
+                "reply_to": reply_to,
                 "subject_line": subject,
             },
         }
 
         if recipients is not None:
-            body["recipients"] = {
+            recipients_mapping: dict[str, Any] = {
                 "list_id": recipients.audience,
             }
 
             if recipients.segment:
                 segment_id = int(recipients.segment.split("/")[1])
-                body["recipients"]["segment_opts"] = {
+                recipients_mapping["segment_opts"] = {
                     "saved_segment_id": segment_id,
                 }
+            body["recipients"] = recipients_mapping
 
         return body
 
@@ -199,9 +224,16 @@ class MailchimpCampaignBackend(CampaignBackend):
         campaign_id: Optional[str] = None,
         recipients: Optional[NewsletterRecipientsBase],
         subject: str,
+        reply_to: str,
+        from_name: str,
         html: str,
     ) -> str:
-        body = self.get_campaign_request_body(recipients=recipients, subject=subject)
+        body = self.get_campaign_request_body(
+            recipients=recipients,
+            subject=subject,
+            from_name=from_name,
+            reply_to=reply_to,
+        )
 
         if campaign_id:
             campaign_id = self._update_campaign(campaign_id, body)
@@ -255,7 +287,7 @@ class MailchimpCampaignBackend(CampaignBackend):
                 error, "Error while sending campaign", campaign_id=campaign_id
             )
 
-    def schedule_campaign(self, campaign_id: str, schedule_time: datetime) -> None:
+    def validate_schedule_time(self, schedule_time: datetime) -> None:
         rounded_minute = schedule_time.minute - (schedule_time.minute % 15)
         rounded_time = schedule_time.replace(
             minute=rounded_minute, second=0, microsecond=0
@@ -266,6 +298,9 @@ class MailchimpCampaignBackend(CampaignBackend):
                 f" e.g. {schedule_time.hour:02d}:{rounded_minute:02d} "
                 f"not {schedule_time.hour:02d}:{schedule_time.minute:02d}."
             )
+
+    def schedule_campaign(self, campaign_id: str, schedule_time: datetime) -> None:
+        self.validate_schedule_time(schedule_time)
 
         try:
             self.client.campaigns.schedule(