+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2025-11-24
+
+### Added
+
+- `NewsletterPageMixin`: add methods to get `from_name` and `reply_to` (@chigby, #92)
+
+## [0.2.3] - 2025-08-20
+
+### Added
+- Validate schedule time before saving to Mailchimp (#82)
+- Mailchimp: Request 1000 audience segments and select only "saved" segments (#84)
+- Development tooling with `uv` (#85)
+- Support for Django 5.2, Wagtail 7.0 and 7.1. (#85)
+- Add extra context to `get_newsletter_html` (#88)
+
+### Removed
+
+- Support for Python 3.9, Django 5.0, Wagtail 5.2 and 6.4. (#85)
+
 ## [0.2.2] - 2025-04-16
 
 ### Added